+++ scripts/start_new_localnet.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-echo "Launching the localnet"
-
-mxpy testnet clean
-mxpy testnet config
-mxpy testnet start